% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMAS_WSC.R
\docType{data}
\name{AMAS_WSC}
\alias{AMAS_WSC}
\title{AMAS Dataset}
\format{
AMAS_WSC is a data frame with 2200 cases (rows) and 9 variables (columns).
The variables are named item_1, item_2, ... , item_9.
}
\source{
Keller et al., ....
}
\usage{
data(AMAS_WSC)
}
\description{
This dataset contains math anxiety measurements using the Abbreviated Math Anxiety Scale (AMAS)
on 2,200 Amazon Mechanical Turk workers (Keller et al., ....)
This scale has 9 items representing mathematics anxiety which is a negative emotion associated
to mathematics related activities.
}
\examples{
summary(AMAS_WSC)
}
\references{
Hopko, D., Mahadevan, R., Bare, R. L., & Hunt, M. K. (2003). The abbreviated math
anxiety scale (AMAS): Construction, validity, and reliability. Assessment, 10 , 78–182.
}
\keyword{datasets}
