% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMASdata.R
\docType{data}
\name{AMASdata}
\alias{AMASdata}
\title{AMAS Dataset}
\format{
amas_wsc is a data frame with 2200 cases (rows) and 9 variables (columns).
The variables are named item1, item2, ... , item9.
}
\source{
Keller et al., ....
}
\usage{
data(AMASdata)
}
\description{
This dataset contains math anxiety measurements using the Abbreviated Math Anxiety Scale (AMAS)
on 2,200 Amazon Mechanical Turk workers (Keller et al., ....)
This scale has 9 items representing mathematics anxiety which is a negative emotion associated
to mathematics related activities.
}
\examples{
summary(amas_wsc)
}
\references{
Hopko, D., Mahadevan, R., Bare, R. L., & Hunt, M. K. (2003). The abbreviated math
anxiety scale (AMAS): Construction, validity, and reliability. Assessment, 10 , 78–182.
}
\keyword{datasets}
