% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDI_WSC.R
\docType{data}
\name{BDI_WSC}
\alias{BDI_WSC}
\title{BDI Dataset}
\format{
BDI_WSC is a data frame with 2200 cases (rows) and 13 variables (columns).
The variables are named item1, item2, ... , item13.
}
\source{
Keller et al., ....
}
\usage{
data(BDI_WSC)
}
\description{
The dataset contains measurements on attitudes and symptoms of depression using the
13-item short form of Beck Depression Inventory (BDI) on 2,200 Amazon Mechanical Turk
workers (Keller et al., ...)
}
\examples{
summary(BDI_WSC)
}
\references{
Beck, A. T., & Beck, R. W. (1972). Screening Depressed Patients in Family
Practice: A Rapid Technic. Postgraduate Medicine, 52(6), 81–85.
https://doi.org/10.1080/00325481.1972.11713319
}
\keyword{datasets}
