% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSES_WSC.R
\docType{data}
\name{GSES_WSC}
\alias{GSES_WSC}
\title{GSES Dataset}
\format{
GSES_WSC is a data frame with 2200 cases (rows) and 6 variables (columns).
The variables are named item1, item2, ... , item 6.
}
\source{
Keller et al., ....
}
\usage{
data(GSES_WSC)
}
\description{
The dataset contains measurements on perceived self-efficacy regarding coping and
adaptation abilities using six out of 10 items of General Self-Efficacy Scale
(GSES, item 2, 3, 6, 5, 7, 10 from the original instrument) on 2,200
Amazon Mechanical Turk workers (Keller et al., ...).
}
\examples{
summary(GSES_WSC)
}
\references{
Schwarzer, R., & Jerusalem, M. (1995). General Self-Efficacy Scale (GSE) \link{Database record}. APA PsycTests.
https://doi.org/10.1037/t00393-000
}
\keyword{datasets}
