% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math_Post_WSC.R
\docType{data}
\name{Math_Post_WSC}
\alias{Math_Post_WSC}
\title{Math Post-test Dataset}
\format{
Math_Post_WSC is a data frame with 2200 cases (rows) and 15 variables (columns).
The variables are named item_1, item_2, ... , item_15.
}
\source{
Keller et al., ....
}
\usage{
data(Math_Post_WSC)
}
\description{
This dataset contains mathematics posttest prepared by the authors (Keller et al., ....) for
the 2,200 Amazon Mechanical Turk workers,
This test has 15 items: five focused on general mathematics aptitude
and ten items dealing with properties of exponents.
}
\examples{
summary(Math_Post_WSC)
}
\references{
Keller et al., ....
}
\keyword{datasets}
