% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math_Pre_WSC.R
\docType{data}
\name{Math_Pre_WSC}
\alias{Math_Pre_WSC}
\title{Math Pre-test Dataset}
\format{
Math_Pre_WSC is a data frame with 2200 cases (rows) and 12 variables (columns).
The variables are named item_1, item_2, ... , item_12.
}
\source{
Keller et al., ....
}
\usage{
data(Math_Pre_WSC)
}
\description{
This dataset contains mathematics pretest prepared by the authors (Keller et al., ....) for
the 2,200 Amazon Mechanical Turk workers,
drawing on a combination of self-written items and items from the ETS Kit of Factor-Referenced
Cognitive Tests.
This test has 12 items: seven items target general mathematics aptitude,
and five items focus on simplifying algebraic expres- sions involving exponents.
}
\examples{
summary(Math_Pre_WSC)
}
\references{
R. B. Ekstrom, J. W. French, H. H. Harman, and D. Dermen. Manual for Kit of
Factor- Referenced Cognitive Tests. Educational Testing Service, 1976.
}
\keyword{datasets}
