% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSC_WSC.R
\docType{data}
\name{MSC_WSC}
\alias{MSC_WSC}
\title{MSC Dataset}
\format{
MSC_WSC is a data frame with 2200 cases (rows) and 7 variables (columns).
The variables are named item1, item2 ...., iteme6, and type.
\itemize{
\item I understand the meaning of a factor in an algebraic expression.
\item I understand the difference between a base and an expression in an algebraic expression.
\item I can simplify algebraic expressions that involve multiplying exponential expressions with the same base.
\item I can simplify algebraic expressions that involve dividing exponential expressions with the same base.
\item I can simplify algebraic expressions that involve raising an expression to a power.
\item I can simplify algebraic expressions that involve negative exponents.
}
}
\source{
Keller et al., ....
}
\usage{
data(MSC_WSC)
}
\description{
This dataset contains measurements of confidence in understanding and ability to
simplify algebraic expressions involving exponents. The 5-point Likert-type items
were developed by Keller et al. (...) and administered to 2,200 Amazon Mechanical
Turk workers. The dataset includes Math Confidence Scale (MCS) measurements taken before
and after a math training intervention. The final variable, named 'type', indicates whether
the data were collected before or after the intervention.
}
\examples{
summary(MSC_WSC)
}
\references{
Keller et al., ....
}
\keyword{datasets}
