% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Big5_WSC.R
\docType{data}
\name{Big5_WSC}
\alias{Big5_WSC}
\title{Big 5 Dataset}
\format{
Big5_WSC is a data frame with 2200 cases (rows) and 44 variables (columns).
The variables are named O1-O10, C1-C9, E1-E8, A1-A9, and N1-N8.
}
\usage{
data(Big5_WSC)
}
\description{
This dataset contains measurements of the Big Five personality trait measurements using the Big Five Inventory (BFI),
collected from 2,200 Amazon Mechanical Turk workers (Keller et al., 2022). The items include both
positively and negatively worded statements; therefore, reverse coding is required
for certain items prior to conducting any psychometric analyses.
}
\details{
The Big Five personality test comprises five scales: Openness (O) with 10 items,
Conscientiousness (C) with 9 items, Extraversion (E) with 8 items, Agreeableness (A)
with 9 items, and Neuroticism (N) with 8 items.
}
\examples{
Big5_WSC
}
\references{
John, O. P., & Srivastava, S. (1999). The big five trait taxonomy: History, measurement,
and theoretical perspectives. In L. A. Pervin & O. P. John (Eds.), Handbook of
personality: Theory and research (pp. 102-138). New York, NY: Guilford Press.
}
\keyword{datasets}
