% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSES_WSC.R
\docType{data}
\name{GSES_WSC}
\alias{GSES_WSC}
\title{GSES Dataset}
\format{
GSES_WSC is a data frame with 2200 cases (rows) and 6 variables (columns).
The variables are named item_1, item_2, ... , item_6.
}
\usage{
data(GSES_WSC)
}
\description{
The dataset contains measurements of perceived self-efficacy related to coping and
adaptation abilities, using six out of the ten items from the General Self-Efficacy Scale
(GSES)-specifically item 2, 3, 6, 5, 7, and 10. The data were collected from 2,200
Amazon Mechanical Turk workers (Keller et al., 2022).
}
\examples{
GSES_WSC
}
\references{
Schwarzer, R., & Jerusalem, M. (1995). General Self-Efficacy Scale (GSE) Database record. APA PsycTests.
https://doi.org/10.1037/t00393-000
}
\keyword{datasets}
