% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math_Post_WSC.R
\docType{data}
\name{Math_Post_WSC}
\alias{Math_Post_WSC}
\title{Math Post-test Dataset}
\format{
Math_Post_WSC is a data frame with 2200 cases (rows) and 15 variables (columns).
The variables are named item_1, item_2, ... , item_15.
}
\usage{
data(Math_Post_WSC)
}
\description{
This dataset contains mathematics posttest data prepared by the authors (Keller et al., 2022), collected from
the 2,200 Amazon Mechanical Turk workers,
This test consists of 15 items: five assessing general mathematics aptitude
and ten focused on properties of exponents.
}
\examples{
Math_Post_WSC
}
\keyword{datasets}
