% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math_Train_WSC.R
\docType{data}
\name{Math_Train_WSC}
\alias{Math_Train_WSC}
\title{Math Training Dataset}
\format{
Math_Train_WSC is a data frame with 2200 cases (rows) and 10 variables (columns).
The variables are named item_1, item_2, ... , item_10.
}
\usage{
data(Math_Train_WSC)
}
\description{
This dataset contains practice questions presented during mathematics intervention
to 2,200 Amazon Mechanical Turk workers.
A total 10 items assessed participants' understanding of rules for simplifying exponential expressions.
For participants who received the vocabulary intervention, all items are recorded as missing.
}
\examples{
Math_Train_WSC
}
\keyword{datasets}
