% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vocab_Post_WSC.R
\docType{data}
\name{Vocab_Post_WSC}
\alias{Vocab_Post_WSC}
\title{Vocabulary Post-test Dataset}
\format{
Vocab_Post_WSC is a data frame with 2200 cases (rows) and 24 variables (columns).
The variables are named item_1, item_2, ... , item_24.
}
\usage{
data(Vocab_Post_WSC)
}
\description{
This dataset contains vocabulary posttest for the 2,200 Amazon Mechanical Turk workers.
This test consists of 24 items that ask participants to match words with their definitions.
Twelve of the fifty words presented during the vocabulary
study session also appeared on the 24-item vocabulary posttest.
}
\examples{
Vocab_Post_WSC
}
\keyword{datasets}
