% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vocab_Pre_WSC.R
\docType{data}
\name{Vocab_Pre_WSC}
\alias{Vocab_Pre_WSC}
\title{Vocabulary Pre-test Dataset}
\format{
Vocab_Pre_WSC is a data frame with 2200 cases (rows) and 24 variables (columns).
The variables are named item_1, item_2, ... , item_24.
}
\usage{
data(Vocab_Pre_WSC)
}
\description{
This dataset contains vocabulary pretest data for 2,200 Amazon Mechanical Turk workers.
The test consists of 24 items that ask participants to match words with their definitions.
}
\examples{
Vocab_Pre_WSC
}
\keyword{datasets}
